pyproject.toml
[project]
name = "jobfinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cloudscraper>=1.2.71",
    "langchain>=1.1.3",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-redis>=0.2.1",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
]
